Fully implementation of D3 with refines for those review comments. (#273)
* feature, implementation of d3, autonity misbehaviour detector.

* no lint, no lint errors for upstream's test.

* refine, refine the variables, better comments for functions of D3.

* resovle the rebase error of merging develop to d3 branch.

* resolve the rebase error for the embedded selector db.

* todo, address root cause of 4byte test failure, add todo in comments.

* add comment for the bug address in the 4bytes_test.

* resolve comments, resolve review comments from dtebbs.

* resolve review comments, better typing for rules and event type.

* refine the fault simulator of d3's integration test.

* replace the fault simulator of PN rule in new test framework.

* resolve review comment, fix typos for d3.

* resolve view comments, replace d3's integration tests with new framework.

* fix lint error.

* ci, split malicious tests from coverage tests in an individual job.

* fix typos.

* ci, fix typo in ci configuraiton.

* refine, remove useless variable.

* resolve review comment, rm unnecessary condition checker.

* improvement, resolve d3 review comments of DoS attack.

* refine, to resolve review comments for d3.
  1. refined capped buffer for future height msg.
  2. do equivocation checking before proposal checker.
  3. go style loops.
  4. message payload rename to bytes.
  5. refine msg store.
  6. refine the ruleEngineLoop() and msgHandlerLoop() in case of potential dead lock.
  7. some other minor improvements.

* refine, move the getProposer buffer to the AC wrapper.

* refine, rename the upstream's ErrFutureBlock to ErrFutureTimestampBlock.

* refine, resolve review comments from team.
  1. put just quorum votes as the proof of accountability event.
  2. rm a sub condition checker for the misbehaviour of PO rule.
  3. refine the MsgHash() of consensus message.
  4. DeEquivocatedMsg() is replaced by function OverQuorumVotes()

* refine, remove GetValidRound() from consensus msg interface.

* refine, make partially signed garbage msg accountable in D3.

* refine, involve only quorum votes as the innocence proof of Accusation PO.

* test, testcases for accountable garbage msg testing in d3.

* improvement, accountability event reporter with slot and TX chunk management.

* test, add chunked accountability event tests.

* improvement, test and refine for the chunked accountability event reporter.

* refine, remove math.abs(r-rPrime) since rPrime is less than r in the context.

* improvement, make chunked event processing more gas economical.

* todo, add comment to refine the accountability reporter.

* refine, remove the Accusation of C since it could be covered by C1.

* refine, check reporter of accountability event in contract.

* refine, replace powerOfVotes with OverQuorumVotes.

* test, tests, refine and bug correction for afd event reporter.

* improvement, resolve dynamic chunk size for oversize afd event.

* improvement, add ruleID checker in AC for the future penalty protocol.

* test, increase the timeout of tbft misbehaviour patterns test.

* refine, reduce the nesting of ifs in rule engine to improve readability.

* fix, include nil prevote in the prevotes set of vr in old prevote accountability checker.

* fix, missing rule for invalid old proposal and prevotes for invalid old proposal.

* improvement, resolve review comments.
  1. merge functions of msg store gc.
  2. use etherBase in accountability event issuer.
  3. use constant accountability event handler method.

* improvement, move accountability event reporter to fault detector package.

* The protocol to process off-chain accusation and innocence proofs. (#220)

improvement, off chain accusation protocol.

* fix, flush trie node for at least accountability height range before exiting.

* lint, fix lint error.

* improvement, panic when exec accountability event over corrupted trieDB client.

* improve, move the range check to contract side to support replaying of TX.

* refine, refine the return code of precompiled contract.

* fix, remove range check in precomile contract since it was move to solidity side.

* doc, update comments in the code.

* refine, just flush AFD prefered num of tries on chain stop, leave TrieInMemory unchanged.

* improve, update comments and extend the flush range of dirty trie on node stop.

* fix, refine contract and fix the legacy panic in msg.validate().

* doc, add todos and align the afd event range in client and ac.

* fix, do the rewind range check by taking the block number from evm context.

* fix, fix legacy tests and remove out of updated comments.

* fix, fix test, AFD contract gas cost weighted by input data size.

* feature, add bad proposal info data structure.

* fix, check if message is nil in the proof of accountability event to improve the robustness.

* refine, remove useless checking at decoding of accountability proof.

* feature, proposal accountability with lite proposal mechanism.

* test, fix unit tests for lite proposal proof mechanism.

* test, fix legacy tendermint tests for lite proposal mechanism.

* lint, no lints for the test which uses rand.Intn().

* feature, make no deps to state db from accountability contract.

* comment, remove useless comments.

* test, fix dynamic rpc test.

* test, add unit test for lite proposal message.

* test, add unit test for getProposer()

* fix, skip equivocated proposal on invalid proposal accounting.

* doc, rm out of updated todos.

* fix, convert protposal to lite one at prevote accountability rules.

* fix todo, take only quorum prevotes as proof for PVN accusation.

* fix, resolve todos in rule engine.

* fix, resolve the review findings on misbehaviour proof of PVN rule.

* fix, resolve review finding that garbage msg proof shouldn't carry evidences.

* fix, remove the accountability over msg of none committee members.

* fix, fix the incorrect len checker in the proof of PVO misbehaviour.

* fix, refine the checker in validMisbehaviourOfPVO for better understanding.

* fix, limit the accountable garbage msg throughout committee members.

* refine, don't save garbage proof to the external nodes anymore at contract.

* fix, allow nil prevote as evidence for PVO rule in contract.

* improve, add redundant checker in case of abuse of proof.

* improvement, check the evidence msg does not exceed the MaxEvidenceMessages.

* improvement, check maxEvidenceMessages as well for innocence proof msg.

* improvement, add old proposal type checker of PO accusation and innocence proof.

* improvement, to resolve review comments.

* improvement, fix lint and remove useless checker for future msg handling.

* fix, take corresponding proposal as the proof of PVN accusation.

* refine, remove invalid step rule since it's duplicated with garbage rule.

* refine, replace the off chain accusation buffer with mapping.

* refine, refine the evidence checker in the innocence proof of PVN accusation.

* improvement, check off-chain accusation exists before handle the innocence proof.

* lint, fix lint.

* ci, fix contract binder generator in Makefile.

* test, fix legacy tests after the rebase.

* test, fix legacy tests for the rebase.

* ci, skip unstable TestTBFTMisbehaviourRulePN due to the simulation is not stable.

* Merge PVO1 and PVO2 in a single rule (#258)

* merge pvo1 and pvo2

* fix misbehaviour detector tests

* fix e2e test

* fix evidence length check

* add naming clarification

* fix evidence lenggth checkl

* fix unit test

---------


* lint, fix lint error.

* improvement, to resolve review comments.

* improvement, to resolve review comments.

* fix, fix test.

* improvement, to prevent from panic on node shutdown.

* test, skip unstable test case.

* fix, merge a correction from d4 during d4's testing.

---------

+++ consensus/tendermint/core/types/timeout.go
@@ -21,7 +21,7 @@ type TimeoutEvent struct {
 	RoundWhenCalled  int64
 	HeightWhenCalled *big.Int
 	// message type: MsgProposal MsgPrevote	MsgPrecommit
-	Step uint64
+	Step uint8
 }
 
 type Timeout struct {
